poc_jpommier/requirements.txt
apachelogs==0.6.0
click==8.1.7
psycopg2-binary==2.9.9
